IroIro

+++ CieloEstrellado/Robot_APIs/PALineSensor.cpp
@@ -1,6 +1,6 @@
 #include "PALineSensor.h"
 
-PALineSensor::PALineSensor(PinName _pin, float _blackValue) : pin(_pin), blackValue(_blackValue)
+PALineSensor::PALineSensor(PinName _pin, uint16_t _blackValue) : pin(_pin), blackValue(_blackValue)
 {
     
 }
